backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PPTX Processing
python-pptx==0.6.23
Pillow==10.2.0

# PDF Generation
reportlab==4.0.8
pikepdf==8.11.2
pypdf==3.17.4

# AI Integration
anthropic==0.18.1

# Language Detection
langdetect==1.0.9

# Utilities
aiofiles==23.2.1
python-magic==0.4.27
uuid6==2024.1.12

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
